pyproject.toml
[project]
name = "ResNet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "prettytable>=3.17.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tabulate>=0.10.0",
    "tensorboard>=2.20.0",
    "thop>=0.1.1.post2209072238",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130"},
]
torchvision = [
  { index = "pytorch-cu130"},
]
torchaudio = [
  { index = "pytorch-cu130"},
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
